refactor(restore): rename and split worker files
